Refactor the modules structure into the components and move the export the epics and reducers from the index duck
