+++ client/js/components/ping/ping-component.jsx
