+++ client/js/components/shared/error-block-component.jsx
