--- client/js/ducks/epics.js
@@ -1,6 +1,0 @@
-import { combineEpics } from 'redux-observable';
-import { pingEpic } from './ping';
-import { rootEpic } from './root';
-import { starwarsEpic } from './starwars';
-
-export default combineEpics(pingEpic, starwarsEpic, rootEpic);
