--- client/js/ducks/reducer.js
@@ -1,10 +1,0 @@
-import { combineReducers } from 'redux';
-import ping from './ping';
-import root from './root';
-import starwars from './starwars';
-
-export default combineReducers({
-    root,
-    ping,
-    starwars
-});
