+++ client/js/router/index.js
@@ -1,13 +1,12 @@
-import Ping from 'modules/root/ping/ping-container';
+import GitHub from 'components/github/github-container';
+import Ping from 'components/ping/ping-container';
 import React from 'react';
-import Root from 'modules/root/root-container';
 import { Route } from 'react-router';
-import StarWars from 'modules/root/starwars/starwars-container';
+import StarWars from 'components/starwars/starwars-container';
 
 export default
 <div>
-    <Route path="/" exact render={ () => <h1>Root</h1> }></Route>
-    <Route path="/github" component={Root}></Route>
+    <Route exact path="/" component={GitHub}></Route>
     <Route path="/ping" component={Ping}></Route>
     <Route path="/starwars" component={StarWars}></Route>
 </div>;
